add solution 3rd task

+++ task3/Program.cs
@@ -1,7 +1,7 @@
 ﻿//Напишите программу, которая на вход принимает число и выдаёт, является ли число чётным (делится ли оно на два без остатка).
 
 Console.WriteLine("Введите число = ");
-double A = double.Parse(Console.ReadLine()!);
+int A = int.Parse(Console.ReadLine()!);
 
 if( A % 2 == 0 )
 {
